chore: refactor graph tools and some queries that do not belong in the data pipeline
