+++ packages/go/dawgs/cardinality/roaring64.go
@@ -1,17 +1,17 @@
 // Copyright 2023 Specter Ops, Inc.
-// 
+//
 // Licensed under the Apache License, Version 2.0
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-// 
+//
 // SPDX-License-Identifier: Apache-2.0
 
 package cardinality
@@ -80,10 +80,26 @@ func (s bitmap64) Remove(value uint64) {
 	s.bitmap.Remove(value)
 }
 
+func (s bitmap64) Xor(provider Provider[uint64]) {
+	switch typedProvider := provider.(type) {
+	case bitmap64:
+		s.bitmap.Xor(typedProvider.bitmap)
+
+	case Duplex[uint64]:
+		providerCopy := roaring64.New()
+
+		typedProvider.Each(func(value uint64) (bool, error) {
+			providerCopy.Add(value)
+			return true, nil
+		})
+
+		s.bitmap.Xor(providerCopy)
+	}
+}
 func (s bitmap64) And(provider Provider[uint64]) {
 	switch typedProvider := provider.(type) {
 	case bitmap64:
-		s.bitmap.Or(typedProvider.bitmap)
+		s.bitmap.And(typedProvider.bitmap)
 
 	case Duplex[uint64]:
 		s.Each(func(nextValue uint64) (bool, error) {
